refactored code. review comments incorporated
