+++ graph_snyk_cve_sync.py
@@ -18,117 +18,21 @@
 """Script which synchronizes snyk CVEs from S3 to graph."""
 
 from datetime import datetime, timedelta
-import boto3
 import json
 import os
 from f8a_utils.versions import get_versions_and_latest_for_ep
 from f8a_utils.golang_utils import GolangUtils
 from unified_range import api
 from f8a_version_comparator.comparable_version import ComparableVersion
+from helper import Helper
 import re
 import logging
-import requests
 import time
 
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 SUPPORTED_ECOSYSTEMS = ['maven', 'npm', 'pypi', 'golang']
-
-
-class Helper:
-    """Helper class with some utility functions."""
-
-    def __init__(self):
-        """Init method for helper class."""
-        self.CVE_BUCKET = os.environ.get("REPORT_BUCKET_NAME", '')
-        self.AWS_KEY = os.environ.get("AWS_S3_ACCESS_KEY_ID_REPORT_BUCKET", '')
-        self.AWS_SECRET = os.environ.get("AWS_S3_SECRET_ACCESS_KEY_REPORT_BUCKET", '')
-        self.AWS_REGION = os.environ.get("AWS_S3_REGION", "us-east-1")
-        self.HOST = os.environ.get('BAYESIAN_DATA_IMPORTER_SERVICE_HOST', 'bayesian-data-importer')
-        self.PORT = os.environ.get('BAYESIAN_DATA_IMPORTER_SERVICE_PORT', '9192')
-
-        self.s3_resource = boto3.resource('s3', aws_access_key_id=self.AWS_KEY,
-                                          aws_secret_access_key=self.AWS_SECRET,
-                                          region_name=self.AWS_REGION)
-
-    def store_json_content(self, content, obj_key):
-        """Store the report content to the S3 storage."""
-        try:
-            logger.info('Storing the data into the S3 file %s' % obj_key)
-            self.s3_resource.Object(self.CVE_BUCKET, obj_key).put(
-                Body=json.dumps(content, indent=2).encode('utf-8'))
-        except Exception as e:
-            logger.exception('%r' % e)
-
-    def _retrieve_dict(self, object_key):
-        """Retrieve a dictionary stored as JSON from S3."""
-        return json.loads(self._retrieve_blob(object_key).decode('utf-8'))
-
-    def _retrieve_blob(self, object_key):
-        """Retrieve remote object content."""
-        return self.s3_resource.Object(self.CVE_BUCKET, object_key).get()['Body'].read()
-
-    def read_data_from_s3(self, date, loc):
-        """Read the snyk data from S3."""
-        try:
-            filename = loc + date + ".json"
-            logger.info('Retrieving the data from the S3 file %s' % filename)
-            return self._retrieve_dict(filename)
-        except Exception as e:
-            logger.error(e)
-            return False
-
-    def make_api_call(self, payload, mode):
-        """Make an API call to data importer."""
-        try:
-            api_url = "http://" + self.HOST + ":" + self.PORT
-            headers = {'Content-type': 'application/json'}
-            msg = ""
-            response = ""
-            if mode == "PUT":
-                response = requests.put('{}/api/v1/snyk-cves'.format(api_url), headers=headers,
-                                        data=json.dumps(payload))
-                msg = "ingested"
-                key = payload['package']
-            elif mode == "DELETE":
-                response = requests.delete('{}/api/v1/snyk-cves'.format(api_url), headers=headers,
-                                           data=json.dumps(payload))
-                msg = "deleted"
-                key = payload['id']
-            if response.status_code == 200:
-                logger.info("CVEs for {p} successfully {m}".format(p=key, m=msg))
-                return "success"
-            elif response.status_code == 504:
-                logger.info("Operation for {p} is taking time. "
-                            "Will be {m} in the background".format(p=key, m=msg))
-                return "delayed"
-            else:
-                logger.info("Error while {m} CVEs for {p}".format(p=key, m=msg))
-                logger.info("Status Code {}".format(response.status_code))
-            return "failed"
-        except Exception:
-            logger.error("API call to data importer failed.")
-
-    def is_dry_run(self):
-        """Return True if this is a dry run."""
-        # Set this value to true if you want the entire operation to run, but not the ingestion.
-        return os.environ.get('SNYK_DRY_RUN', 'false').lower() in ('1', 'yes', 'true')
-
-    def force_run_ingestion(self):
-        """Return if ingestion mode is on."""
-        # Set this value when you want to run the ingestion forcefully (ignores runtime).
-        return os.environ.get('SNYK_INGESTION_FORCE_RUN', 'false').lower() in ('1', 'yes', 'true')
-
-    def is_delta_mode_on(self):
-        """Return if the delta feed mode is on."""
-        # Set this value if you want to run only in the diff mode.
-        return os.environ.get('SNYK_DELTA_FEED_MODE', 'false').lower() in ('1', 'yes', 'true')
-
-    def ingestion_run_time(self):
-        """Return the time when ingestion needs to run."""
-        # Set the time at which you want the ingestion to run.
-        return os.environ.get('SNYK_INGESTION_RUN_TIME', '12')
 
 
 class SnykCveSync:
@@ -147,7 +51,6 @@ class SnykCveSync:
         # This is the offset days i.e how many days old data do we need to ingest.
         delta_feed_offset = int(os.environ.get('SNYK_DELTA_FEED_OFFSET', '1'))
         self.start_day = self.today - timedelta(days=delta_feed_offset)
-        self.selective_eco_run = os.environ.get('SELECTIVE_ECOSYSTEM_SNYK_SYNC', 'none')
 
         # If we want to selectively run the ingestion for an ecosystem.
         self.selective_eco_run = os.environ.get('SELECTIVE_ECOSYSTEM_SNYK_SYNC', '')
@@ -277,15 +180,19 @@ class SnykCveSync:
         for eco in self.snyk_data:
             if eco == "java" and self._parse_data_for_eco(eco):
                 logger.info("Parsing feed for Maven.")
+                self._add_default_obj_for_eco("maven")
                 self._parse_data(self.snyk_data[eco], "maven")
             elif eco == "js" and self._parse_data_for_eco(eco):
                 logger.info("Parsing feed for Npm.")
+                self._add_default_obj_for_eco("npm")
                 self._parse_data(self.snyk_data[eco], "npm")
             elif eco == "python" and self._parse_data_for_eco(eco):
                 logger.info("Parsing feed for Pypi.")
+                self._add_default_obj_for_eco("pypi")
                 self._parse_data(self.snyk_data[eco], "pypi")
             elif eco == "golang" and self._parse_data_for_eco(eco):
                 logger.info("Parsing feed for Golang.")
+                self._add_default_obj_for_eco("golang")
                 self._parse_golang_data(self.snyk_data[eco], "golang")
             else:
                 logger.info("Ignoring the ecosystem {} from the feed".format(eco))
@@ -296,6 +203,22 @@ class SnykCveSync:
         """When running under delta feed mode, we need to consider only those vulns which were
         updated between the given offset date and today's date."""
         return self.today > date_obj >= self.start_day
+
+    def _add_default_obj_for_eco(self, eco):
+        """Add default entries for the ecosystem into the different lists."""
+        self.CVE_DATA[eco] = {}
+        self.DELETE_CVE_DATA[eco] = []
+        self.DELTA_FEED[eco] = []
+
+    def _add_data_for_false_positives(self, eco, data, pkg):
+        """Update the records with false positive data."""
+        self.DELETE_CVE_DATA[eco].append(data)
+        self.DELTA_FEED[eco].append(data)
+        # Default status is skipped. When ingested, it gets updated with success or failed.
+        self.SNYK_REPORT['details'][eco]['delete'][data['id']] = {
+            'name': pkg,
+            'status': "skipped"
+        }
 
     def _generate_default_cve_obj(self, eco, pkg, versions, latest, gh=None, lic=None):
         """Generate the default cve object."""
@@ -319,14 +242,22 @@ class SnykCveSync:
             obj['license'] = lic
         return obj
 
+    def _set_additional_fields(self, data):
+        """To set additional fields or modify some values before ingesting."""
+        # Remove the non required rules data.
+        del data['rules']
+        # Change description into proper string.
+        data['description'] = re.sub("[\'\"]", "", data['description'])
+        # Calculate and update the premium field.
+        premium = str(data.get('premium', "false")).lower() == 'true'
+        data['pvtVuln'] = premium
+        return data
+
     def _parse_golang_data(self, vuln_data, eco):
         """Parse data for golang eco."""
         total_vuln = 0
         delta_mode = self.helper.is_delta_mode_on()
         if len(vuln_data) != 0:
-            self.CVE_DATA[eco] = {}
-            self.DELETE_CVE_DATA[eco] = []
-            self.DELTA_FEED[eco] = []
             for data in vuln_data:
                 # If delta mode is on & the modificationTime doesnt fall in the range, then ignore.
                 if delta_mode and not self._is_date_in_range(data['modificationTime']):
@@ -338,12 +269,7 @@ class SnykCveSync:
                     if len(data['vulnerableVersions']) == 0:
                         # In this case, we use the data to remove the vuln from the garph
                         logger.info("False positive found. {i}".format(i=data['id']))
-                        self.DELETE_CVE_DATA[eco].append(data)
-                        self.DELTA_FEED[eco].append(data)
-                        self.SNYK_REPORT['details'][eco]['delete'][data['id']] = {
-                            'name': pkg,
-                            'status': "skipped"
-                        }
+                        self._add_data_for_false_positives(eco, data, pkg)
                         continue
                     """ This is done so that we dont fetch the same pkg data again & again if more
                     than 1 vuln for the same pkg is present."""
@@ -385,16 +311,13 @@ class SnykCveSync:
                             commit hash usecase, even when affected versions not found"""
                             logger.info("No affected versions for {}".format(data['id']))
                         total_vuln += 1
-                        del data['rules']
-                        data['description'] = re.sub("[\'\"]", "", data['description'])
-                        premium = str(data.get('premium', "false")).lower() == 'true'
-                        data['pvtVuln'] = premium
+                        data = self._set_additional_fields(data)
                         # In Snyk feed, for some golang vuln, they dont have this field.
                         if 'vulnerableHashes' not in data:
                             data['vulnerableHashes'] = []
                         self.SNYK_REPORT['details'][eco]['ingest'][data['id']] = {
                             'name': pkg,
-                            'premium': premium,
+                            'premium': data['pvtVuln'],
                             'affected_version_count': len(data['affected']),
                             'affected_commit_hash_count': len(data['vulnerableHashes']),
                             'status': "skipped"
@@ -417,9 +340,6 @@ class SnykCveSync:
         total_vuln = 0
         delta_mode = self.helper.is_delta_mode_on()
         if len(vuln_data) != 0:
-            self.CVE_DATA[eco] = {}
-            self.DELETE_CVE_DATA[eco] = []
-            self.DELTA_FEED[eco] = []
             for data in vuln_data:
                 if eco == "pypi":
                     data['package'] = str.lower(data['package'])
@@ -433,12 +353,7 @@ class SnykCveSync:
                     if len(data['vulnerableVersions']) == 0:
                         # In this case, we use the data to remove the vuln from the garph
                         logger.info("False positive found. {i}".format(i=data['id']))
-                        self.DELETE_CVE_DATA[eco].append(data)
-                        self.DELTA_FEED[eco].append(data)
-                        self.SNYK_REPORT['details'][eco]['delete'][data['id']] = {
-                            'name': pkg,
-                            'status': "skipped"
-                        }
+                        self._add_data_for_false_positives(eco, data, pkg)
                         continue
                     """ This is done so that we dont fetch the same pkg data again & again if more
                     than 1 vuln for the same pkg is present."""
@@ -471,16 +386,13 @@ class SnykCveSync:
                                         .format(pkg))
                             continue
                         total_vuln += 1
-                        del data['rules']
                         self.CVE_DATA[eco][pkg]['affected'].extend(data['affected'])
                         self.CVE_DATA[eco][pkg]['affected'] = list(
                             set(self.CVE_DATA[eco][pkg]['affected']))
-                        data['description'] = re.sub("[\'\"]", "", data['description'])
-                        premium = str(data.get('premium', "false")).lower() == 'true'
-                        data['pvtVuln'] = premium
+                        data = self._set_additional_fields(data)
                         # Update the premium value and the affected version count in report.
                         rep_obj = self.SNYK_REPORT['details'][eco]['ingest'][data['id']]
-                        rep_obj['premium'] = premium
+                        rep_obj['premium'] = data['pvtVuln']
                         rep_obj['affected_version_count'] = len(data['affected'])
                         self.DELTA_FEED[eco].append(data)
                         self.CVE_DATA[eco][pkg]['vulnerabilities'].append(data)
