+++ snyk_feed.py
@@ -24,7 +24,8 @@ import requests
 import logging
 import json
 from datetime import datetime
-from graph_snyk_cve_sync import Helper, SnykCveSync
+from graph_snyk_cve_sync import SnykCveSync
+from helper import Helper
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
