+++ tests/test_graph_snyk_cve_sync.py
@@ -5,8 +5,8 @@ from unittest.mock import patch
 from datetime import datetime
 import os
 from f8a_version_comparator.comparable_version import ComparableVersion
-
-from graph_snyk_cve_sync import Helper, SnykCveSync
+from helper import Helper
+from graph_snyk_cve_sync import SnykCveSync
 
 data = {
     "ruby": [],
@@ -368,11 +368,11 @@ data = {
 }
 
 scs = SnykCveSync(data)
-h = Helper()
+helper = Helper()
 
 
 @patch("graph_snyk_cve_sync.SnykCveSync._parse_data_for_eco")
-@patch("graph_snyk_cve_sync.Helper.is_delta_mode_on")
+@patch("helper.Helper.is_delta_mode_on")
 def test_extract_data_from_feed(m1, m2):
     """Test extract_data_from_feed."""
     m1.return_value = False
@@ -444,7 +444,7 @@ def test_extract_data_from_feed(m1, m2):
     assert len(report_data['npm']['ingest']) == 3
 
 
-@patch("graph_snyk_cve_sync.Helper.is_delta_mode_on")
+@patch("helper.Helper.is_delta_mode_on")
 def test_extract_data_from_feed1(m1):
     """Test extract_data_from_feed in delta mode."""
     m1.return_value = True
@@ -478,7 +478,7 @@ def test_extract_data_from_feed1(m1):
 
 
 @responses.activate
-@patch('graph_snyk_cve_sync.Helper.is_dry_run')
+@patch('helper.Helper.is_dry_run')
 def test_insert_cves(mocker):
     """Test insert_cves."""
     mocker.return_value = True
@@ -526,7 +526,7 @@ def test_insert_cves(mocker):
 
 
 @responses.activate
-@patch('graph_snyk_cve_sync.Helper.is_dry_run')
+@patch('helper.Helper.is_dry_run')
 def test_delete_cves(mocker):
     """Test _delete_cves."""
     mocker.return_value = True
@@ -616,9 +616,9 @@ def test_delete_cves(mocker):
 @responses.activate
 @patch('graph_snyk_cve_sync.SnykCveSync._generate_snyk_report')
 @patch('graph_snyk_cve_sync.SnykCveSync._delete_cves')
-@patch('graph_snyk_cve_sync.Helper.store_json_content')
+@patch('helper.Helper.store_json_content')
 @patch('graph_snyk_cve_sync.SnykCveSync.disable_snyk_run')
-@patch('graph_snyk_cve_sync.Helper.read_data_from_s3')
+@patch('helper.Helper.read_data_from_s3')
 @patch('graph_snyk_cve_sync.SnykCveSync._extract_data_from_feed')
 @patch('graph_snyk_cve_sync.SnykCveSync._insert_cves')
 def test_run_snyk_sync(m1, m2, m3, m4, m5, m6, m7):
@@ -651,7 +651,7 @@ def test_disable_snyk_run():
     assert val is True
 
 
-@patch('graph_snyk_cve_sync.Helper.force_run_ingestion')
+@patch('helper.Helper.force_run_ingestion')
 def test_disable_snyk_run1(m1):
     """Test disable_snyk_run1."""
     m1.return_value = True
@@ -662,46 +662,46 @@ def test_disable_snyk_run1(m1):
 @patch.dict(os.environ, {'SNYK_DRY_RUN': 'true'})
 def test_is_dry_run():
     """Test is_dry_run."""
-    val = h.is_dry_run()
+    val = helper.is_dry_run()
     assert val is True
 
 
 @patch.dict(os.environ, {'SNYK_DRY_RUN': '0'})
 def test_is_dry_run1():
     """Test is_dry_run."""
-    val = h.is_dry_run()
+    val = helper.is_dry_run()
     assert val is False
 
 
 @patch.dict(os.environ, {'SNYK_INGESTION_FORCE_RUN': 'true'})
 def test_force_run_ingestion():
     """Test force_run_ingestion."""
-    val = h.force_run_ingestion()
+    val = helper.force_run_ingestion()
     assert val is True
 
 
 @patch.dict(os.environ, {'SNYK_INGESTION_FORCE_RUN': '0'})
 def test_force_run_ingestion1():
     """Test force_run_ingestion."""
-    val = h.force_run_ingestion()
+    val = helper.force_run_ingestion()
     assert val is False
 
 
 @patch.dict(os.environ, {'SNYK_DELTA_FEED_MODE': 'true'})
 def test_is_delta_mode_on():
     """Test is_delta_mode_on."""
-    val = h.is_delta_mode_on()
+    val = helper.is_delta_mode_on()
     assert val is True
 
 
 @patch.dict(os.environ, {'SNYK_DELTA_FEED_MODE': '0'})
 def test_is_delta_mode_on1():
     """Test is_delta_mode_on."""
-    val = h.is_delta_mode_on()
+    val = helper.is_delta_mode_on()
     assert val is False
 
 
-@patch('graph_snyk_cve_sync.Helper.is_delta_mode_on')
+@patch('helper.Helper.is_delta_mode_on')
 def test_parse_data_for_eco(m1):
     """Test parse_data_for_eco."""
     # Delta mode is on, so run for all ecosystems
