+++ tests/test_snyk_feed.py
@@ -4,9 +4,11 @@ import pytest
 import os
 from unittest.mock import patch
 from datetime import datetime
+from helper import Helper
 
 from snyk_feed import SnykDataFetcher
 sdf = SnykDataFetcher()
+helper = Helper()
 
 
 class S3Obj:
@@ -28,7 +30,7 @@ def test_disable_snyk_run():
     assert val is True
 
 
-@patch('snyk_feed.Helper.force_run_ingestion')
+@patch('helper.Helper.force_run_ingestion')
 def test_disable_snyk_run1(m1):
     """Test disable_snyk_run1."""
     m1.return_value = True
